+++ packages/iov-crypto/types/slip10.d.ts
@@ -37,3 +37,4 @@ export declare class Slip10 {
   private static n;
 }
 export declare function pathToString(path: readonly Slip10RawIndex[]): string;
+export declare function stringToPath(input: string): readonly Slip10RawIndex[];